Major updates for ATB: > Restoration Modes will mostly work once per second > Refactored DC_TurnManager > Buff duration is reduced gradually

+++ C-Engine/src/main/java/main/ability/PassiveAbilityObj.java
@@ -70,8 +70,8 @@ public class PassiveAbilityObj extends AbilityObj implements Attachment {
     }
 
     @Override
-    public int tick() {
-        return 0;
+    public Double tick() {
+        return 0.0;
     }
 
     @Override
